# docs/swagger/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to the server
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error or file not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload a file
...
swagger: "2.0"